+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 
 ### Changed
 - HttpClientRouter now throws a HttpClientNoMatchException instead of a RequestException if it can not find a client for the request.
-- RetryPlugin will no longer retry requests when the response failed with a HTTP code < 500.
+- RetryPlugin will only retry exceptions when there is no response, or a response in the 5xx HTTP code range.
+- RetryPlugin also retries when no exception is thrown if the responses has HTTP code in the 5xx range.
+  The callbacks for exception handling have been renamed and callbacks for response handling have been added.
 - Abstract method `HttpClientPool::chooseHttpClient()` has now an explicit return type (`Http\Client\Common\HttpClientPoolItem`)
 - Interface method `Plugin::handleRequest(...)` has now an explicit return type (`Http\Promise\Promise`)
 - Made  classes final that are not intended to be extended.
@@ -16,6 +18,7 @@
 
 ### Removed
 - Deprecated option `debug_plugins` has been removed from `PluginClient`
+- Deprecated options `decider` and `delay` have been removed from `RetryPlugin`, use `exception_decider` and `exception_delay` instead.
 
 ## 1.9.0 - 2019-01-03